Implement delete context menu action for ImageViewerFragment
